requirements.txt
fastapi==0.112.2
starlette==0.37.2
uvicorn==0.30.5
psycopg[binary]==3.2.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.9.2
python-dotenv==1.2.1
